+++ src/babashka/bbin/specs.clj
@@ -4,6 +4,4 @@
 (s/def ::ns-default symbol?)
 (s/def ::main-opts (s/coll-of string?))
 (s/def ::script-config (s/keys :opt-un [::main-opts ::ns-default]))
-(s/def ::bin (s/map-of symbol? ::script-config))
-(s/def :bbin/bin ::bin)
-(s/def ::bbin (s/keys :opt-un [::bin]))
+(s/def :bbin/bin (s/map-of symbol? ::script-config))